Patch by Alexandro D. Almeida: * Cast the slider value to double before dividing by 1000 in order to apply   a smooth speed value. Fixes #1470. Thanks!
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@26866 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/preferences/mouse/MouseWindow.cpp
@@ -161,7 +161,7 @@ MouseWindow::MessageReceived(BMessage *message)
 			int32 value;
 			if (message->FindInt32("be:value", &value) == B_OK) {
 				// slow = 8192, fast = 524287
-				fSettings.SetMouseSpeed((int32)pow(2, value * 6 / 1000) * 8192);
+				fSettings.SetMouseSpeed((int32)pow(2, (double)value * 6 / 1000) * 8192);
 				fDefaultsButton->SetEnabled(fSettings.IsDefaultable());
 				fRevertButton->SetEnabled(true);
 			}
